requirements.txt
opencv-python-headless
numpy
matplotlib
pillow
streamlit